backend/python/requirements.txt
# Flask核心
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6

# 数据库
psycopg2-binary==2.9.9
redis==5.0.1

# 密码加密
bcrypt==4.1.2
cryptography==41.0.8

# 数据验证
marshmallow==3.20.2
email-validator==2.1.0

# HTTP客户端
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# 日期时间处理
python-dateutil==2.8.2

# 图片处理
Pillow==10.1.0

# 文件上传
Flask-Uploads==0.2.1

# 任务队列
celery==5.3.4
flower==2.0.1

# 监控和日志
sentry-sdk[flask]==1.38.0
prometheus-flask-exporter==0.23.0

# 开发工具
python-dotenv==1.0.0
gunicorn==21.2.0

# 测试
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# 代码质量
flake8==6.1.0
black==23.11.0
isort==5.12.0

# API文档
flasgger==0.9.7.1

# 缓存
Flask-Caching==2.1.0

# 邮件
Flask-Mail==0.9.1

# 定时任务
APScheduler==3.10.4
